refactor: move Pane struct to pane.rs
