+++ pretty/src/lib.rs
@@ -24,6 +24,6 @@ pub use self::notation::{
     Notation, Repeat,
 };
 
-pub use self::pretty::{Pane, PlainText, PrettyDocument, PrettyWindow};
+pub use self::pretty::{PlainText, PrettyDocument, PrettyWindow};
 pub use self::style::{Color, ColorTheme, Emph, Rgb, Shade, Style};
-pub use pane::{render_pane, Content, PaneError, PaneNotation, PaneSize};
+pub use pane::{render_pane, Content, Pane, PaneError, PaneNotation, PaneSize};
