+++ pretty/src/pretty/mod.rs
@@ -4,4 +4,4 @@ mod pretty_window;
 
 pub use plain_text::PlainText;
 pub use pretty_doc::PrettyDocument;
-pub use pretty_window::{Pane, PrettyWindow};
+pub use pretty_window::PrettyWindow;
